move renumber command to a separate module.
