+++ scripts/tests/bindings.lua
@@ -34,7 +34,6 @@ function test_process()
 	assert_equal("function", _G.type(eressea.process.siege))
 	assert_equal("function", _G.type(eressea.process.leave))
 	assert_equal("function", _G.type(eressea.process.promote))
-	assert_equal("function", _G.type(eressea.process.renumber))
 	assert_equal("function", _G.type(eressea.process.restack))
 	assert_equal("function", _G.type(eressea.process.set_spells))
 	assert_equal("function", _G.type(eressea.process.set_help))
